Move all the server CLI stuff under `kubetools server`.
