+++ kubetools_client/cli/__init__.py
@@ -58,15 +58,7 @@ class SpecialHelpOrder(click.Group):
 
 @click.group(cls=SpecialHelpOrder)
 @click.option(
-    '-s', '--server',
-    envvar='KUBETOOLS_HOST',
 )
-@click.option(
-    '-p', '--port',
-    type=int,
-    envvar='KUBETOOLS_PORT',
-)
-@click.option('kube_env', '--env', '--kube-env')
 @click.option('--debug', is_flag=True)
 @click.version_option(version=__version__, message='%(prog)s: v%(version)s')
 @click.pass_context
@@ -75,15 +67,7 @@ def cli_bootstrap(ctx, server=None, port=None, kube_env=None, debug=False):
     Kubetools client.
     '''
 
+
+
     setup_logging(debug)
-
-    # Get settings
-    settings = get_settings(debug)
-
-    # CLI options > settings
-    host = server or settings.KUBETOOLS_HOST
-    port = port or settings.KUBETOOLS_PORT
-    kube_env = kube_env or settings.DEFAULT_KUBE_ENV
-
-    # Setup the client
-    ctx.obj = KubeClient(host=host, port=port, kube_env=kube_env)
+    get_settings(debug)  # trigger settings load
