+++ kubetools_client/cli/__main__.py
@@ -3,7 +3,11 @@
 from kubetools_client.cli import cli_bootstrap
 from kubetools_client.main import run_cli
 # Import click command groups
-from kubetools_client.cli import configs, generate, jobs, lists, wait  # noqa: F401, I100
+from kubetools_client.cli import (  # noqa: F401, I100
+    configs,
+    deploy,
+    server,
+)
 
 
 run_cli(cli_bootstrap)
