+++ kubetools_client/cli/server.py
@@ -1,14 +1,15 @@
-# Kubetools Client
-# File: kubetools/cli/__init__.py
-# Desc: helpers for the Kubetools command line interface
+import sys
 
 import click
 import six
 
+from kubetools_client import KubeClient
 from kubetools_client.cli import cli_bootstrap
-from kubetools_client.cli.wait import wait_build, wait_job
 from kubetools_client.exceptions import KubeCLIError
 from kubetools_client.log import logger
+from kubetools_client.settings import get_settings
+
+from .server_util import wait_with_spinner
 
 
 def _get_apps_in_namespace(client, namespace):
@@ -111,7 +112,187 @@ def _check_existing_apps(client, namespace, action):
         ))
 
 
-@cli_bootstrap.command()
+def wait_job(client, namespace, job_id):
+    def check_job_status(previous_status):
+        namespace_jobs = client.list_jobs_by_namespace(namespace=namespace)
+        job = namespace_jobs[namespace][job_id]
+
+        if (
+            # Job completed the number of times we want?
+            job['succeeded'] == job['completions']
+            # Job has failed over X times?
+            or job['failed'] >= job['backoffLimit']
+        ):
+            return
+
+        return job['annotations'].get('description', 'Running job')
+
+    wait_with_spinner(check_job_status)
+
+    # Get the job again
+    namespace_jobs = client.list_jobs_by_namespace(namespace=namespace)
+    job = namespace_jobs[namespace][job_id]
+    status = job['succeeded'] == job['completions']
+
+    if status:
+        click.echo('<-- Job complete, final status = {0}'.format(
+            click.style('SUCCESS', 'green'),
+        ))
+
+    else:
+        click.echo('<-- Job complete, final status = {0}'.format(
+            click.style('ERROR', 'red'),
+        ))
+
+    if not status:
+        sys.exit(1)
+
+
+def wait_build(client, build_hash):
+    '''
+    CLI helper that waits for a build to complete while displaying a little spinner.
+    '''
+
+    def check_build_status(previous_status):
+        build_data = client.get_build(build_hash)
+
+        # If we're complete, just return nothing
+        if build_data['status'] not in ('PENDING', 'RUNNING'):
+            return
+
+        status_text = build_data['status']
+        if build_data['status_detail']:
+            status_text = '{0}/{1}'.format(status_text, build_data['status_detail'])
+
+        return status_text
+
+    wait_with_spinner(check_build_status)
+
+    # Get the build date
+    build_data = client.get_build(build_hash)
+    status = build_data['status']
+
+    # We did it!
+    if status == 'SUCCESS':
+        click.echo('--> Namespace: {0}'.format(click.style(
+            build_data['namespace'], bold=True,
+        )))
+
+        domains = build_data.get('domains')
+        if domains:
+            for app, domain in domains.items():
+                click.echo('--> App: {0} should now be available @ {1}'.format(
+                    app, click.style(domain, bold=True),
+                ))
+
+    elif status == 'ABORTED':
+        click.echo('--> {0}'.format(
+            click.style(
+                'Aborted at stage: {0}'.format(build_data['status_detail']),
+                'yellow',
+            ),
+        ))
+
+    # Status should be ERROR, but any non-success here is a fail
+    else:
+        click.echo('--> {0}'.format(
+            click.style(
+                'Error at stage: {0}'.format(build_data['status_detail']),
+                'red',
+            ),
+        ))
+        click.echo()
+        click.echo('--------ERROR START--------')
+        click.echo(click.style(build_data.get('error', 'Unknown error'), 'yellow'))
+        click.echo('---------ERROR END---------')
+
+    if status == 'SUCCESS':
+        formatted_status = click.style('SUCCESS', 'green')
+    elif status == 'ABORTED':
+        formatted_status = click.style('ABORTED', 'yellow')
+    else:
+        formatted_status = click.style('ERROR', 'red')
+
+    click.echo('<-- Build complete, final status = {0}'.format(formatted_status))
+
+    if status == 'ABORTED':
+        sys.exit(1)
+
+    elif status != 'SUCCESS':
+        sys.exit(2)
+
+
+@cli_bootstrap.group(help_priority=2)
+@click.option(
+    '-s', '--server',
+    envvar='KUBETOOLS_HOST',
+)
+@click.option(
+    '-p', '--port',
+    type=int,
+    envvar='KUBETOOLS_PORT',
+)
+@click.option('kube_env', '--env', '--kube-env')
+@click.pass_context
+def server(ctx, server=None, port=None, kube_env=None):
+    '''
+    Trigger builds on a Kubetools server.
+    '''
+
+    settings = get_settings()
+
+    # CLI options > settings
+    host = server or settings.KUBETOOLS_HOST
+    port = port or settings.KUBETOOLS_PORT
+    kube_env = kube_env or settings.DEFAULT_KUBE_ENV
+
+    # Setup the client
+    ctx.obj = KubeClient(host=host, port=port, kube_env=kube_env)
+
+
+@server.command()
+@click.argument('namespace')
+@click.argument('job_id')
+@click.pass_obj
+def abort_job(client, namespace, job_id):
+    if client.abort_job(namespace, job_id):
+        click.echo('--> Job aborted: {0}/{1}'.format(
+            namespace,
+            click.style(job_id, bold=True),
+        ))
+    else:
+        click.echo(click.style('Failed to abort job!', 'red'))
+
+
+@server.command()
+@click.argument('namespace')
+@click.argument('job_id')
+@click.pass_obj
+def delete_job(client, namespace, job_id):
+    if client.delete_job(namespace, job_id):
+        click.echo('--> Job deleted: {0}/{1}'.format(
+            namespace,
+            click.style(job_id, bold=True),
+        ))
+    else:
+        click.echo(click.style('Failed to delete job!', 'red'))
+
+
+@server.command()
+@click.argument('build_hash')
+@click.pass_obj
+def wait(client, build_hash):
+    '''
+    Wait for a Kubetools build.
+    '''
+
+    click.echo('### Watching build: {0}'.format(
+        click.style(build_hash, bold=True),
+    ))
+    wait_build(client, build_hash)
+
+
+@server.command()
 @click.argument('build_options', nargs=-1, required=True)
 @click.pass_obj
 def build(client, build_options):
@@ -124,7 +305,7 @@ def build(client, build_options):
     )
 
 
-@cli_bootstrap.command()
+@server.command()
 @click.argument('namespace')
 @click.argument('deploy_options', nargs=-1, required=True)
 @click.option(
@@ -146,7 +327,7 @@ def deploy(client, deploy_options, namespace=None, envar=None):
     )
 
 
-@cli_bootstrap.command()
+@server.command()
 @click.argument('namespace')
 @click.argument('upgrade_options', nargs=-1)
 @click.pass_obj
@@ -164,7 +345,7 @@ def upgrade(client, namespace, upgrade_options):
     )
 
 
-@cli_bootstrap.command()
+@server.command()
 @click.argument('namespace')
 @click.argument('apps_to_remove', nargs=-1)
 @click.pass_obj
@@ -182,7 +363,7 @@ def remove(client, namespace, apps_to_remove):
     )
 
 
-@cli_bootstrap.command()
+@server.command()
 @click.argument('namespace')
 @click.argument('apps_to_restart', nargs=-1)
 @click.pass_obj
@@ -200,7 +381,7 @@ def restart(client, namespace, apps_to_restart):
     )
 
 
-@cli_bootstrap.command()
+@server.command()
 @click.argument('namespace')
 @click.pass_obj
 def cleanup(client, namespace):
@@ -216,7 +397,7 @@ def cleanup(client, namespace):
     )
 
 
-@cli_bootstrap.command()
+@server.command()
 @click.argument('namespace')
 @click.argument('app')
 @click.argument('command')
